chore: rename loginsignup pages
