+++ src/router/index.js
@@ -2,7 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import ProductDetailPage from '../views/ProductDetailPage.vue'
 import CartView from '@/views/CartView.vue'
 import HomeView from '@/views/HomeView.vue'
-import LoginSignup from '@/views/LoginSignup.vue'
+import AuthForm from '@/views/AuthForm.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -70,9 +70,9 @@ const router = createRouter({
       component: () => import('@/views/OrderManagement.vue'),
     },
     {
-      path: '/loginsignup',
-      name: 'loginsignup',
-      component: LoginSignup,
+      path: '/authform',
+      name: 'authform',
+      component: AuthForm,
     },
   ],
 })
